Fix for recording proxy sessions
When a proxy is in use, Mojo::UserAgent will create a new HTTP CONNECT
transaction, create a copy of the transaction that initiated the HTTP
CONNECT transaction, and then process that copy. MUA::Mockable will no
longer record the HTTP CONNECT transaction or the aforementioned copy.
In playback mode, it will be as if there was no proxy at all.